Cache logic. Copy and move element.

+++ lemon-tree/admin/src/lib/FileStore.php
@@ -1,10 +1,11 @@
 <?php namespace LemonTree;
 
-use Illuminate\Cache\StoreInterface;
+use Exception;
 use Illuminate\Cache\TaggableStore;
 use Illuminate\Filesystem\Filesystem;
+use Illuminate\Contracts\Cache\Store;
 
-class CustomFileStore extends TaggableStore implements StoreInterface {
+class FileStore extends TaggableStore implements Store {
 
 	/**
 	 * The Illuminate Filesystem instance.
@@ -41,6 +42,17 @@ class CustomFileStore extends TaggableStore implements StoreInterface {
 	 */
 	public function get($key)
 	{
+		return array_get($this->getPayload($key), 'data');
+	}
+
+	/**
+	 * Retrieve an item and expiry time from the cache by key.
+	 *
+	 * @param  string  $key
+	 * @return array
+	 */
+	protected function getPayload($key)
+	{
 		$path = $this->path($key);
 
 		// If the file doesn't exists, we obviously can't return the cache so we will
@@ -48,16 +60,16 @@ class CustomFileStore extends TaggableStore implements StoreInterface {
 		// the expiration UNIX timestamps from the start of the file's contents.
 		if ( ! $this->files->exists($path))
 		{
-			return null;
+			return array('data' => null, 'time' => null);
 		}
 
 		try
 		{
 			$expire = substr($contents = $this->files->get($path), 0, 10);
 		}
-		catch (\Exception $e)
+		catch (Exception $e)
 		{
-			return null;
+			return array('data' => null, 'time' => null);
 		}
 
 		// If the current time is greater than expiration timestamps we will delete
@@ -65,10 +77,19 @@ class CustomFileStore extends TaggableStore implements StoreInterface {
 		// this directory much cleaner for us as old files aren't hanging out.
 		if (time() >= $expire)
 		{
-			return $this->forget($key);
+			$this->forget($key);
+
+			return array('data' => null, 'time' => null);
 		}
 
-		return unserialize(substr($contents, 10));
+		$data = unserialize(substr($contents, 10));
+
+		// Next, we'll extract the number of minutes that are remaining for a cache
+		// so that we can properly retain the time for things like the increment
+		// operation that may be performed on the cache. We'll round this out.
+		$time = ceil(($expire - time()) / 60);
+
+		return compact('data', 'time');
 	}
 
 	/**
@@ -100,7 +121,7 @@ class CustomFileStore extends TaggableStore implements StoreInterface {
 		{
 			$this->files->makeDirectory(dirname($path), 0777, true, true);
 		}
-		catch (\Exception $e)
+		catch (Exception $e)
 		{
 			//
 		}
@@ -111,13 +132,17 @@ class CustomFileStore extends TaggableStore implements StoreInterface {
 	 *
 	 * @param  string  $key
 	 * @param  mixed   $value
-	 * @return void
-	 *
-	 * @throws \LogicException
+	 * @return int
 	 */
 	public function increment($key, $value = 1)
 	{
-		throw new \LogicException("Increment operations not supported by this driver.");
+		$raw = $this->getPayload($key);
+
+		$int = ((int) $raw['data']) + $value;
+
+		$this->put($key, $int, (int) $raw['time']);
+
+		return $int;
 	}
 
 	/**
@@ -125,13 +150,11 @@ class CustomFileStore extends TaggableStore implements StoreInterface {
 	 *
 	 * @param  string  $key
 	 * @param  mixed   $value
-	 * @return void
-	 *
-	 * @throws \LogicException
+	 * @return int
 	 */
 	public function decrement($key, $value = 1)
 	{
-		throw new \LogicException("Decrement operations not supported by this driver.");
+		return $this->increment($key, $value * -1);
 	}
 
 	/**
@@ -150,7 +173,7 @@ class CustomFileStore extends TaggableStore implements StoreInterface {
 	 * Remove an item from the cache.
 	 *
 	 * @param  string  $key
-	 * @return void
+	 * @return bool
 	 */
 	public function forget($key)
 	{
@@ -158,8 +181,10 @@ class CustomFileStore extends TaggableStore implements StoreInterface {
 
 		if ($this->files->exists($file))
 		{
-			$this->files->delete($file);
+			return $this->files->delete($file);
 		}
+
+		return false;
 	}
 
 	/**
@@ -169,9 +194,12 @@ class CustomFileStore extends TaggableStore implements StoreInterface {
 	 */
 	public function flush()
 	{
-		foreach ($this->files->directories($this->directory) as $directory)
+		if ($this->files->isDirectory($this->directory))
 		{
-			$this->files->deleteDirectory($directory);
+			foreach ($this->files->directories($this->directory) as $directory)
+			{
+				$this->files->deleteDirectory($directory);
+			}
 		}
 	}
